importar docentes por excel

+++ Backend/Rutas/Importacion_exportacion.py
@@ -1,6 +1,6 @@
 from Extensiones import db
 from flask import Blueprint, jsonify,request
-from Funciones.ImportarExcel import importar_grupos,importar_calificaciones
+from Funciones.ImportarExcel import importar_grupos,importar_calificaciones, importar_docentes
 from Modelos.Modelos import Grupos, Carreras,Materias
 import os
 from Funciones.Registrar_moviminto import registrar_audi
@@ -78,6 +78,43 @@ def importar_Excel_Calificaciones():
         archivo.save(ruta)
 
         procesamiento =  importar_calificaciones(ruta,grupos_nombres,materias_nombres)
+       
+
+        os.remove(ruta)
+
+        if procesamiento.startswith("No se pudo importar") or "Error" in procesamiento:
+            return jsonify({'error': procesamiento}), 400
+        
+        registrar_audi('Alumnos','Importar grupo',212420)
+
+        return jsonify({'mensaje':'Archivo subido', 'resultado' : procesamiento})
+    except Exception  as e:
+        return jsonify({'error': f'Error al importar el archivo {str(e)}'}),400
+
+
+@Import_export_bp.route('/api/importar/ddocentes',methods=['POST'])
+def importar_Excel_Calificaciones():
+    try:
+        UPLOAD_FOLDER = 'docentes'
+        os.makedirs(UPLOAD_FOLDER, exist_ok=True)
+        
+        #validaciones 
+        if 'archivo' not in request.files:
+            return jsonify({'error': 'No fue enviaddo un archivo'}),400
+        
+        archivo = request.files['archivo']
+
+        if archivo.filename == '':
+            return jsonify({'error': 'El nombre del archivo está vacío'}),400
+        
+        if not archivo.filename.endswith(('.xls','.xlsx')):
+            return jsonify({'error':'Formato de archivo no válido'}),400
+        
+
+        ruta = os.path.join(UPLOAD_FOLDER, archivo.filename)
+        archivo.save(ruta)
+
+        procesamiento =  importar_docentes(ruta)
        
 
         os.remove(ruta)
